Eduardo api BD Create

+++ dentalnet-backend/app/Http/Controllers/Admin/Rol/RolesController.php
@@ -13,7 +13,10 @@ class RolesController extends Controller
      */
     public function index(Request $request)
     {
-        // Filtro por nombre de rol
+        // QUE EL FILTRO POR NOMBRE DE ROL
+        if(!auth('api')->user()->can('list_rol')){
+            return response()->json(["message" => "EL USUARIO NO ESTA AUTORIZADO"],403);
+        }
         $name = $request->search;
 
         $roles = Role::where("name","like","%".$name."%")->orderBy("id","desc")->get();
@@ -21,11 +24,11 @@ class RolesController extends Controller
         return response()->json([
             "roles" => $roles->map(function($rol) {
                 return [
-                    "id" => $rol,
+                    "id" => $rol->id,
                     "name" => $rol->name,
-                    "permision" => $rol->permission,
+                    "permision" => $rol->permissions,
+                    "permision_pluck" => $rol->permissions->pluck("name"),
                     "created_at" => $rol->created_at->format("Y-m-d h:i:s")
-
                 ];
             }),
         ]);
@@ -36,12 +39,15 @@ class RolesController extends Controller
      */
     public function store(Request $request)
     {
-        $is_role = Role::where("name", $request->name)->first();
+        if(!auth('api')->user()->can('register_rol')){
+            return response()->json(["message" => "EL USUARIO NO ESTA AUTORIZADO"],403);
+        }
+        $is_role = Role::where("name",$request->name)->first();
 
         if($is_role){
             return response()->json([
                 "message" => 403,
-                "message_text" => "El nombre del rol ya existe"
+                "message_text" => "EL NOMBRE DEL ROL YA EXISTE"
             ]);
         }
 
@@ -50,8 +56,8 @@ class RolesController extends Controller
             'name' => $request->name,
         ]);
         // ["register_rol","edit_rol","register_paciente"];
-        foreach ($permisions as $key => $permision) {
-            $role->givePermissionTo('permision');
+        foreach ($request->permisions as $key => $permision) {
+            $role->givePermissionTo($permision);
         }
         return response()->json([
             "message" => 200,
@@ -63,7 +69,18 @@ class RolesController extends Controller
      */
     public function show(string $id)
     {
-        //
+        if(!auth('api')->user()->can('edit_rol')){
+            return response()->json(["message" => "EL USUARIO NO ESTA AUTORIZADO"],403);
+        }
+        $role = Role::findOrFail($id);
+        return response()->json([
+            "id" => $role->id,
+            "name" => $role->name,
+            "permision" => $role->permissions,
+            "permision_pluck" => $role->permissions->pluck("name"),
+            "created_at" => $role->created_at->format("Y-m-d h:i:s")
+        ]);
+        
     }
 
     /**
@@ -71,19 +88,22 @@ class RolesController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $is_role = Role::where("id","<>",$id)->where("name", $request->name)->first();
- 
+        if(!auth('api')->user()->can('edit_rol')){
+            return response()->json(["message" => "EL USUARIO NO ESTA AUTORIZADO"],403);
+        }
+        $is_role = Role::where("id","<>",$id)->where("name",$request->name)->first();
+
         if($is_role){
             return response()->json([
                 "message" => 403,
-                "message_text" => "El nombre del rol ya existe"
+                "message_text" => "EL NOMBRE DEL ROL YA EXISTE"
             ]);
         }
 
         $role = Role::findOrFail($id);
 
         $role->update($request->all());
-        //["register_rol","edit_rol","register_paciente"];
+        // ["register_rol","edit_rol","register_paciente"];
         $role->syncPermissions($request->permisions);
         return response()->json([
             "message" => 200,
@@ -95,7 +115,16 @@ class RolesController extends Controller
      */
     public function destroy(string $id)
     {
+        if(!auth('api')->user()->can('delete_rol')){
+            return response()->json(["message" => "EL USUARIO NO ESTA AUTORIZADO"],403);
+        }
         $role = Role::findOrFail($id);
+        if($role->users->count() > 0){
+            return response()->json([
+                "message" => 403,
+                "message_text" => "EL ROL SELECCIONADO NO SE PUEDE ELIMINAR POR MOTIVOS QUE YA TIENE USUARIOS RELACIONADOS"
+            ]);
+        }
         $role->delete();
         return response()->json([
             "message" => 200,
