+++ dentalnet-backend/app/Http/Controllers/Admin/Staff/StaffsController.php
@@ -2,7 +2,10 @@
 
 namespace App\Http\Controllers\Admin\Staff;
 
+use Carbon\Carbon;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Spatie\Permission\Models\Role;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Storage;
@@ -16,13 +19,18 @@ class StaffsController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny',User::class);
+
         $search = $request->search;
 
-        $users = User::where("name","like","%".$search."%")
-                        ->orWhere("surname","like","%".$search."%")
-                        ->orWhere("email","like","%".$search."%")
-                        ->orWhere("surname","like","%".$search."%")
+        $users = User::where(DB::raw("CONCAT(users.name,' ',IFNULL(users.surname,''),' ',users.email)"),"like","%".$search."%")
+                        // "name","like","%".$search."%"
+                        // ->orWhere("surname","like","%".$search."%")
+                        // ->orWhere("email","like","%".$search."%")
                         ->orderBy("id","desc")
+                        ->whereHas("roles",function($q){
+                            $q->where("name","not like","%DOCTOR%");
+                        })
                         ->get();
 
         return response()->json([
@@ -31,7 +39,7 @@ class StaffsController extends Controller
     }
 
     public function config() {
-        $roles = Role::all();
+        $roles = Role::where("name","not like","%DOCTOR%")->get();
 
         return response()->json([
             "roles" => $roles
@@ -42,13 +50,13 @@ class StaffsController extends Controller
      */
     public function store(Request $request)
     {
-        
+        $this->authorize('create',User::class);
         $users_is_valid = User::where("email",$request->email)->first();
 
         if($users_is_valid){
             return response()->json([
                 "message" => 403,
-                "message_text" => "El usuario con este email ya existe"
+                "message_text" => "EL USUARIO CON ESTE EMAIL YA EXISTE"
             ]);
         }
 
@@ -60,6 +68,11 @@ class StaffsController extends Controller
         if($request->password){
             $request->request->add(["password" => bcrypt($request->password)]);
         }
+        // "Fri Oct 08 1993 00:00:00 GMT-0500 (hora estándar de Perú)"
+        // Eliminar la parte de la zona horaria (GMT-0500 y entre paréntesis)
+        $date_clean = preg_replace('/\(.*\)|[A-Z]{3}-\d{4}/', '', $request->birth_date);
+
+        $request->request->add(["birth_date" => Carbon::parse($date_clean)->format("Y-m-d h:i:s")]);
 
         $user = User::create($request->all());
 
@@ -76,10 +89,11 @@ class StaffsController extends Controller
      */
     public function show(string $id)
     {
+        $this->authorize('view',User::class);
         $user = User::findOrFail($id);
 
         return response()->json([
-            "user" => UserResource::make($user),
+            "user" => UserResource::make($user), 
         ]);
     }
 
@@ -88,22 +102,22 @@ class StaffsController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('update',User::class);
         $users_is_valid = User::where("id","<>",$id)->where("email",$request->email)->first();
 
         if($users_is_valid){
             return response()->json([
                 "message" => 403,
-                "message_text" => "El usuario con este email ya existe"
+                "message_text" => "EL USUARIO CON ESTE EMAIL YA EXISTE"
             ]);
         }
 
         $user = User::findOrFail($id);
-        
         if($request->hasFile("imagen")){
             if($user->avatar){
                 Storage::delete($user->avatar);
             }
-            $path = Storage::putFile("Staffs",$request->file("imagen"));
+            $path = Storage::putFile("staffs",$request->file("imagen"));
             $request->request->add(["avatar" => $path]);
         }
 
@@ -111,12 +125,18 @@ class StaffsController extends Controller
             $request->request->add(["password" => bcrypt($request->password)]);
         }
 
+        if($request->birth_date){
+            $date_clean = preg_replace('/\(.*\)|[A-Z]{3}-\d{4}/', '', $request->birth_date);
+            $request->request->add(["birth_date" => Carbon::parse($date_clean)->format("Y-m-d h:i:s")]);
+        }
+
+        // $request->request->add(["birth_date" => Carbon::parse($request->birth_date, 'GMT')->format("Y-m-d h:i:s")]);
         $user->update($request->all());
 
-        if($request->role_id != $user->roles()->first()->id){
+        if($request->role_id && $request->role_id != $user->roles()->first()->id){
             $role_old = Role::findOrFail($user->roles()->first()->id);
             $user->removeRole($role_old);
-
+    
             $role_new = Role::findOrFail($request->role_id);
             $user->assignRole($role_new);
         }
@@ -130,6 +150,7 @@ class StaffsController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete',User::class);
         $user = User::findOrFail($id);
         if($user->avatar){
             Storage::delete($user->avatar);
